Add database

+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -1,7 +1,10 @@
 package ru.yandex.practicum.filmorate.controller;
 
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.Positive;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -11,57 +14,73 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.service.UserService;
+import ru.yandex.practicum.filmorate.validator.group.Create;
+import ru.yandex.practicum.filmorate.validator.group.Default;
+import ru.yandex.practicum.filmorate.validator.group.Update;
 
 import java.util.List;
 
 @RestController
+@Validated
 @RequiredArgsConstructor
 @RequestMapping("/users")
 public class UserController {
     private final UserService userService;
 
     @GetMapping
-    public List<User> findAll() {
-        return userService.findAll();
+    public List<User> getUsers() {
+        return userService.getUsers();
     }
 
     @GetMapping("/{id}")
-    public User findById(@PathVariable long id) {
-        return userService.findById(id);
+    public User getUserById(@PathVariable @Positive long id) {
+        return userService.getUserById(id);
     }
 
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
-    public User create(@RequestBody User user) {
-        return userService.create(user);
+    @Validated({Default.class, Create.class})
+    public User create(@RequestBody @Valid User user) {
+        return userService.save(user);
     }
 
     @PutMapping
-    public User update(@RequestBody User newUser) {
+    @Validated({Default.class, Update.class})
+    public User update(@RequestBody @Valid User newUser) {
         return userService.update(newUser);
     }
 
     @PutMapping("/{id}/friends/{friendId}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void addFriend(@PathVariable long id, @PathVariable long friendId) {
+    public void addFriend(@PathVariable @Positive long id, @PathVariable @Positive long friendId) {
+        if (id == friendId) {
+            throw new ValidationException("Пользователь не может добавить самого себя в друзья");
+        }
         userService.addFriend(id, friendId);
     }
 
     @DeleteMapping("/{id}/friends/{friendId}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void removeFriend(@PathVariable long id, @PathVariable long friendId) {
+    public void removeFriend(@PathVariable @Positive long id, @PathVariable @Positive long friendId) {
+        if (id == friendId) {
+            throw new ValidationException("Пользователь не может удалить самого себя из друзей");
+        }
         userService.removeFriend(id, friendId);
     }
 
     @GetMapping("/{id}/friends")
-    public List<User> findAllFriends(@PathVariable long id) {
-        return userService.findAllFriends(id);
+    public List<User> getFriends(@PathVariable @Positive long id) {
+        return userService.getFriends(id);
     }
 
     @GetMapping("/{id}/friends/common/{otherId}")
-    public List<User> findCommonFriends(@PathVariable long id, @PathVariable long otherId) {
-        return userService.findCommonFriends(id, otherId);
+    public List<User> getCommonFriends(@PathVariable @Positive long id, @PathVariable @Positive long otherId) {
+        if (id == otherId) {
+            throw new ValidationException("ID пользователей должны отличаться");
+        }
+        return userService.getCommonFriends(id, otherId);
     }
 }
